# [Fix #10764] Fix performance issue for Layout/FirstHashElementIndentation and Layout/FirstArrayElementIndentation