--- a/changelog/fix_performance_issue_for_multiline_element_indentation.md
+++ b/changelog/fix_performance_issue_for_multiline_element_indentation.md
@@ -1,0 +1,1 @@
+* [#10764](https://github.com/rubocop/rubocop/issues/10764): Fix performance issue for Layout/FirstHashElementIndentation and Layout/FirstArrayElementIndentation. ([@j-miyake][])