Fixes #10764.

In rubocop v1.31.0, `MultilineElementIndentation` that is used by `Layout/FirstHashElementIndentation` and `Layout/FirstArrayElementIndentation` walks all nodes through in the code to find a hash pair node where its value part begins with the given left-brace; That computation is expensive, so I made a change so that the hash pair is passed by each class that uses `MultilineElementIndentation`.

I ran rubocop by the following command to make sure the performance problem was fixed by this PR. The input file is the file provided in https://github.com/rubocop/rubocop/issues/10764#issuecomment-1170011469
```bash
$ time exe/rubocop --cache false test.rb
```

Result:

- With v1.30.1
```
exe/rubocop --cache false test.rb  1.51s user 0.33s system 41% cpu 4.410 total
```

- With v1.31.0
```
exe/rubocop --cache false test.rb  3.26s user 0.29s system 84% cpu 4.199 total
```

- With this PR
```
exe/rubocop --cache false test.rb  1.55s user 0.32s system 62% cpu 3.012 total
```

-----------------

Before submitting the PR make sure the following are checked:

* [x] The PR relates to *only* one subject with a clear title and description in grammatically correct, complete sentences.
* [x] Wrote [good commit messages][1].
* [x] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [x] Squashed related commits together.
* [x] Added tests.
* [x] Ran `bundle exec rake default`. It executes all tests and runs RuboCop on its own code.
* [x] Added an entry (file) to the [changelog folder](https://github.com/rubocop/rubocop/blob/master/changelog/) named `{change_type}_{change_description}.md` if the new code introduces user-observable changes. See [changelog entry format](https://github.com/rubocop/rubocop/blob/master/CONTRIBUTING.md#changelog-entry-format) for details.

[1]: https://chris.beams.io/posts/git-commit/